[HUDI-6432] Fix TestCleanPlanExecutor (#9045)
